Further work on trade system. Buy/Sell information now output to chat so there is a permanent record of it.
Buy/Sell prices now work off the same role, so they are consistent.

Flag to track whether items can be added to an actor has been changed to
the id of the actor they need to be added to. We need to be able to
prevent dragging of items between some actors without first going
through our dialogs.